requirements.txt
streamlit 
openai 
python-dotenv
groq
